+++ packages/components/src/Embed.js
@@ -1,8 +1,18 @@
+// @ts-check
 import React from 'react'
-import Box from './Box'
 
+import Box, { __isBoxStyledSystemProp } from './Box'
+import { getProps } from './util'
+
+const getContainerProps = getProps(__isBoxStyledSystemProp)
+const getIframeProps = getProps((str) => !__isBoxStyledSystemProp(str))
+
+/** @typedef {import("../index").EmbedProps} EmbedProps */
+/** @type {React.ForwardRefExoticComponent<EmbedProps>} */
 export const Embed = React.forwardRef(function Embed(
   {
+    variant,
+    sx,
     ratio = 16 / 9,
     src,
     frameBorder = 0,
@@ -10,20 +20,22 @@ export const Embed = React.forwardRef(function Embed(
     width = 560,
     height = 315,
     allow,
-    ...props
+    ...rest
   },
   ref
 ) {
   return (
     <Box
-      {...props}
+      variant={variant}
+      sx={sx}
       __css={{
         width: '100%',
         height: 0,
         paddingBottom: 100 / ratio + '%',
         position: 'relative',
         overflow: 'hidden',
-      }}>
+      }}
+      {...getContainerProps(rest)}>
       <Box
         ref={ref}
         as="iframe"
@@ -33,6 +45,7 @@ export const Embed = React.forwardRef(function Embed(
         frameBorder={frameBorder}
         allowFullScreen={allowFullScreen}
         allow={allow}
+        {...getIframeProps(rest)}
         __css={{
           position: 'absolute',
           width: '100%',