+++ packages/docs/src/components/debug-theme.js
@@ -2,7 +2,7 @@
 import { jsx, Context } from 'theme-ui'
 import { useContext } from 'react'
 
-export default props => {
+export default (props) => {
   const theme = useContext(Context)
   return <pre children={JSON.stringify(theme, null, 2)} />
 }