+++ packages/docs/src/components/layout.js
@@ -43,7 +43,9 @@ export default (props) => {
   const fullwidth =
     (props.pageContext.frontmatter &&
       props.pageContext.frontmatter.fullwidth) ||
-    props.location.pathname === '/home'
+    props.location.pathname === '/home' ||
+    props.location.pathname === '/home/'
+
   const showNav = !props.pageContext?.frontmatter?.hidenav
 
   const cycleMode = (e) => {