+++ packages/gatsby-theme-ui-layout/src/layout.js
@@ -1,1 +1,1 @@
-export default props => props.children
+export default (props) => props.children