+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## v0.6.0 UNRELEASED
 
+- Add ability for MDX styling, and fix mdx table align styles. Issue #654
 - Remove recursive default values from CSS custom properties. PR #1327
+- Support a `"default"` key for object in scales. PR #951
+- Render extra Embed props onto `iframe` tag instead of wrapping `div`. Issue #966, PR #1122
 
 ## v0.6.0-alpha.1