+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.6.0 UNRELEASED
 
+- Extract objects with nested variant props. Issue #1357
 - Add ability for MDX styling, and fix mdx table align styles. Issue #654
 - Remove recursive default values from CSS custom properties. PR #1327
 - Support a `"default"` key for object in scales. PR #951